ProcessEmailsFunction/requirements.txt
google-auth>=2.0.0
google-api-python-client>=2.0.0
google-cloud-documentai>=2.0.0
google-cloud-storage>=2.0.0
google-cloud-secret-manager>=2.0.0
psycopg2-binary>=2.9.0
functions-framework>=3.0.0
pandas>=1.3.0
python-dotenv>=0.19.0
PyYAML>=6.0
